Fix usage of dynamic variables
The dynamic variables $*SAT-SOLVER, $*SAT-COUNTER and $*SAT-ENUMERATOR
should take precendece over the generic lookup routines in the SAT::
packages in search for a fitting solver.

They have been inserted at the beginning of the sequence of solvers
to test with `$*SAT-SOLVER // Empty`, but the values stored in the
dynamic variables are usually classes which are undefined and always
cause // to return Empty. This is fixed now.